:green_heart: Fixed CI pipeline DB tests updated
