+++ code/load/sql_files/init.sql
@@ -1,4 +1,5 @@
 CREATE DATABASE IF NOT EXISTS MLentory_DB;
+CREATE DATABASE IF NOT EXISTS test_DB;
 USE MLentory_DB;
 
 CREATE TABLE IF NOT EXISTS `Version_Range`(
